fix: 修复 4 个严重代码问题 + 多项架构优化
严重问题修复:
- 资源泄漏: close() 统一为 async def，StockQuoteRepository 加 await
- 性能: AssetQuoteService/AssetVarietyRepository 改为实例字段，避免每次请求重建
- 架构违规: DataSource 层去掉 DB 查询，名称回填上提到 Service 层
- ORM 类型: Numeric(18,4) 列的 Mapped[float] 改为 Mapped[Decimal]

其他优化:
- 清理 _QUERY_ASSET_CLASS 死映射，ETF 直接走 FundQuoteRepository
- 新增子分类 sub_category 字段，ETF 从独立 asset_class 改为 FUND 的子类
- DB 列顺序调整为 id→ticker→name→asset_class→sub_category→market
- AssetVarietyRepository 新增 get_name_map() 批量查询方法
- FundQuoteRepository 内部按 sub_category 自动分流 ETF→腾讯/基金→天天
- API: fund 端点 market 改为路径参数，保持一致

新文件:
- data/dbjson/: DB 品种数据按分类导出
- docs/code_review.md: 代码审查报告（26 项问题跟踪）
- script/merge_etf_fund.py: ETF/FUND 合并脚本

+++ backend/app/models/orm/asset_holding_orm.py
@@ -1,6 +1,7 @@
 """AssetHolding ORM 模型 — 对应 asset_holdings 表"""
 
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import Date, DateTime, Integer, Numeric, String, func
 from sqlalchemy.orm import Mapped, mapped_column
@@ -19,9 +20,9 @@ class AssetHoldingRecord(Base):
     market: Mapped[str] = mapped_column(String(10), nullable=False)
     asset_class: Mapped[str] = mapped_column(String(10), nullable=False)
     currency: Mapped[str] = mapped_column(String(3), default="CNY")
-    quantity: Mapped[float] = mapped_column(Numeric(18, 4), nullable=False)
-    cost_price: Mapped[float] = mapped_column(Numeric(18, 4), nullable=False)
-    total_invested: Mapped[float] = mapped_column(Numeric(18, 4), nullable=False)
+    quantity: Mapped[Decimal] = mapped_column(Numeric(18, 4), nullable=False)
+    cost_price: Mapped[Decimal] = mapped_column(Numeric(18, 4), nullable=False)
+    total_invested: Mapped[Decimal] = mapped_column(Numeric(18, 4), nullable=False)
     first_buy_date: Mapped[datetime] = mapped_column(Date, nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now())
     updated_at: Mapped[datetime | None] = mapped_column(DateTime, onupdate=func.now(), nullable=True)
